Break up the `test_utils.py` file
Having two files with the same name is making mypy unhappy for some reason.
