+++ src/flickypedia/apis/flickr_photos_api/utils.py
@@ -1,78 +1,7 @@
 import datetime
-from typing import Dict, List, Optional
 import xml.etree.ElementTree as ET
 
 from ._types import SafetyLevel, Size, TakenGranularity
-
-
-def find_required_elem(elem: ET.Element, *, path: str) -> ET.Element:
-    """
-    Find the first subelement matching ``path``, or throw if absent.
-    """
-    # We use this when we're parsing responses from the Flickr API, and
-    # there are certain elements we know should always be present in responses.
-    #
-    # e.g. we know that photos always have a <title>, so we could write:
-    #
-    #     photo_elem.find(".//title")
-    #
-    # But the type checker only knows that ``find()`` returns Optional[Element] --
-    # it doesn't know that this path should always be present in the response.
-    #
-    # If we call it from this function instead:
-    #
-    #     find_required(photo_elem, path=".//title")
-    #
-    # Then the type checker can see that it returns a well-defined Element,
-    # and it's happy for us to use it without checking in the rest of the code.
-    matching_elem = elem.find(path=path)
-
-    if matching_elem is None:
-        raise ValueError(f"Could not find required match for {path!r} in {elem!r}")
-
-    return matching_elem
-
-
-def find_required_text(elem: ET.Element, *, path: str) -> str:
-    """
-    Find the text of the first element matching ``path``, or throw if absent.
-    """
-    # We use this when we're parsing responses from the Flickr API, and
-    # there are certain elements we know should always be present and have text.
-    #
-    # e.g. we know that users always have a <id> with some text, so we could write:
-    #
-    #     user_elem.find(".//id")
-    #
-    # But the type checker only knows that ``find()`` returns Optional[Element] --
-    # it doesn't know that this path should always be present in the response.
-    #
-    # If we call it from this function instead:
-    #
-    #     find_required_text(user_elem, path=".//id")
-    #
-    # Then the type checker can see that it returns a well-defined Element,
-    # and it's happy for us to use it without checking in the rest of the code.
-    matching_elem = find_required_elem(elem, path=path)
-    text = matching_elem.text
-
-    if text is None:
-        raise ValueError(f"Could not find required text in {matching_elem}")
-
-    return text
-
-
-def find_optional_text(elem: ET.Element, *, path: str) -> Optional[str]:
-    """
-    Find the text of the first element matching ``path``, or return None
-    if the element is missing, has no text, or has empty text.
-    """
-    matching_elem = elem.find(path=path)
-
-    if matching_elem is None:
-        return None
-
-    return matching_elem.text or None
 
 
 def parse_date_posted(p: str) -> datetime.datetime:
@@ -104,7 +33,7 @@ def parse_date_taken_granularity(g: str) -> TakenGranularity:
 
     See https://www.flickr.com/services/api/misc.dates.html
     """
-    lookup_table: Dict[str, TakenGranularity] = {
+    lookup_table: dict[str, TakenGranularity] = {
         "0": "second",
         "4": "month",
         "6": "year",
@@ -124,7 +53,7 @@ def parse_safety_level(s: str) -> SafetyLevel:
 
     See https://www.flickrhelp.com/hc/en-us/articles/4404064206996-Content-filters
     """
-    lookup_table: Dict[str, SafetyLevel] = {
+    lookup_table: dict[str, SafetyLevel] = {
         "0": "safe",
         "1": "moderate",
         "2": "restricted",
@@ -136,7 +65,7 @@ def parse_safety_level(s: str) -> SafetyLevel:
         raise ValueError(f"Unrecognised safety level: {s}")
 
 
-def parse_sizes(photo_elem: ET.Element) -> List[Size]:
+def parse_sizes(photo_elem: ET.Element) -> list[Size]:
     """
     Get a list of sizes from a photo in a collection response.
     """
@@ -151,7 +80,7 @@ def parse_sizes(photo_elem: ET.Element) -> List[Size]:
     #       …
     #     />
     #
-    sizes: List[Size] = []
+    sizes: list[Size] = []
 
     for suffix, label in [
         ("sq", "Square"),
